+++ python_modules/dagit/dagit/app.py
@@ -1,10 +1,12 @@
-from dagster import check
+from dagster import DagsterInstance, check
+from dagster.cli.workspace.cli_target import get_workspace_process_context_from_kwargs
 from dagster.core.execution.compute_logs import warn_if_compute_logs_disabled
 from dagster.core.telemetry import log_workspace_stats
 from dagster.core.workspace.context import WorkspaceProcessContext
 from starlette.applications import Starlette
 
-from .starlette import DagitWebserver
+from .version import __version__
+from .webserver import DagitWebserver
 
 
 def create_app_from_workspace_process_context(
@@ -32,3 +34,21 @@ def create_app_from_workspace_process_context(
         workspace_process_context,
         path_prefix,
     ).create_asgi_app()
+
+
+def default_app(debug=False):
+    instance = DagsterInstance.get()
+    process_context = get_workspace_process_context_from_kwargs(
+        instance=instance,
+        version=__version__,
+        read_only=False,
+        kwargs={},
+    )
+
+    return DagitWebserver(
+        process_context,
+    ).create_asgi_app(debug=debug)
+
+
+def debug_app():
+    return default_app(debug=True)
