+++ python_modules/dagit/dagit_tests/webserver/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from dagit.starlette import DagitWebserver
+from dagit.webserver import DagitWebserver
 from dagster import DagsterInstance, __version__
 from dagster.cli.workspace.cli_target import get_workspace_process_context_from_kwargs
 from starlette.testclient import TestClient
